auth/requirements.txt
Flask_RESTful==0.3.6
Flask==0.12.2
Flask_Cors==3.0.3
Flask_Script==2.0.6
pymongo==3.5.1
python_bcrypt==0.3.2
requests==2.18.4
